# AUTOMATION/web/ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Soulglade — API du studio",
    "description": "\nBackend local du studio Soulglade. Aucune authentification : le serveur\nn'accepte que ce qui vient de lui-même (voir `api/security.py`).\n\n**Conventions transverses**\n\n* presque toute route lit `?character=<id>` et le valide avant de toucher au\n  disque ; `/img` l'exige (pas de repli sur un personnage par défaut) ;\n* toute réponse porte un corps JSON, succès comme échec. Une erreur a la forme\n  `{\"ok\": false, \"erreur\": \"<texte français destiné à l'écran>\"}` ;\n* les uploads voyagent en base64 dans un corps JSON, jamais en\n  `multipart/form-data` — conséquence voulue du garde d'origine.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/character": {
      "get": {
        "summary": "Fiche du personnage courant",
        "description": "Current character, for the header (registry J4; type + world J7bis) and\nfor its SHEET (F1.2, 30/08/2026).\n\nThe `current_character` dependency already guaranteed the character has a\ncoherent character.json (real pack, (type, style) resolving to it,\ncompatible world) — no extra error handling here.\n\nThe sheet reads everything HERE, in ONE call already bound to the\ncharacter: adding `base` and `nsfw_tool` to this response costs two\nregistry reads, where a second route would have duplicated the pack\nresolution and the isolation that goes with it. Nothing new is computed —\n`edit_tool_state` is the one the Application screen already displays.",
        "operationId": "get_character_api_character_get",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterSheet"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/journal": {
      "get": {
        "summary": "Journal de production du personnage",
        "description": "Production journal, filtered on the requested character.\n\nTruncated to the last 300 rows, newest first, with no pagination — same as\nbefore (AUDIT §7.9). The CSV is a single file for the whole platform; the\nfilter is here, because without it one of Léna's rows illustrates another\ncharacter's image as soon as two file names collide.",
        "operationId": "get_journal_api_journal_get",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gallery": {
      "get": {
        "summary": "Contenu d'un dossier de tri",
        "description": "Content of one sorting folder — of THIS character, and of it alone.\n\n┌── COUPLING TO PRESERVE — migration brief §4.1, AUDIT §5.6.1 ────────────┐\n│ THIS IS WHERE THE `v` TOKEN IS BORN. `\"v\": int(f.stat().st_mtime)` on   │\n│ every item, and nowhere else in the API.                                │\n│                                                                         │\n│ The frontend's single image-URL builder — `imgUrl()` in static/api.js — │\n│ appends it to /img, which ACCEPTS AND IGNORES it (see routers/images.py,│\n│ same box). It is pure cache-busting: since the editor learned to        │\n│ overwrite its source (/api/edit/save?remplacer, F3.3), one file name    │\n│ can designate two different images, and without a changing URL the      │\n│ browser re-serves the old bytes from its cache.                         │\n│                                                                         │\n│ THE FORMAT IS PART OF THE CONTRACT: an INTEGER, the mtime in seconds.   │\n│ Not a float, not an ISO date, not a hash. Emitting it under another     │\n│ name, another type, or dropping it, breaks the cache silently — no      │\n│ error anywhere, just a stale image on screen after an overwrite.        │\n│                                                                         │\n│ Note it is deliberately ABSENT from STATE.recent (routers/production.py)│\n│ — `imgUrl` then omits it and the URL stays character-for-character the  │\n│ one from before. Do not \"harmonise\" that.                               │\n└─────────────────────────────────────────────────────────────────────────┘",
        "operationId": "get_gallery_api_gallery_get",
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "OK",
              "title": "Bucket"
            }
          },
          {
            "name": "space",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Space"
            }
          },
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GalleryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/flag": {
      "post": {
        "summary": "Jugement humain de réalisme",
        "description": "Human judgement on realism. Independent of sorting: it moves nothing.\n\nTook no character parameter at all until 2026-09-01 — the DB write this\ntriggers (`mesures.poser_flag` -> `base.enregistrer_image`) silently\nrecorded every judgement, for every character, under one specific\ncharacter_id, because that was the only default `enregistrer_image` had.",
        "operationId": "set_flag_api_flag_post",
        "parameters": [
          {
            "name": "character",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire.",
              "title": "Character"
            },
            "description": "Identifiant du personnage (registre CHARACTERS/). Obligatoire."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête refusée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterSheet": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "world": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorldBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "output_style": {
            "type": "string",
            "title": "Output Style"
          },
          "universe": {
            "$ref": "#/components/schemas/UniverseBrief"
          },
          "content_types": {
            "additionalProperties": true,
            "type": "object",
            "title": "Content Types"
          },
          "nsfw": {
            "type": "boolean",
            "title": "Nsfw"
          },
          "base": {
            "$ref": "#/components/schemas/FrozenBaseBrief"
          },
          "nsfw_tool": {
            "$ref": "#/components/schemas/EditToolState"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "output_style",
          "universe",
          "nsfw",
          "base",
          "nsfw_tool"
        ],
        "title": "CharacterSheet",
        "description": "The current character, for the chrome and for its sheet (F1.2)."
      },
...
      "ErrorResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": false
          },
          "erreur": {
            "type": "string",
            "title": "Erreur"
          }
        },
        "type": "object",
        "required": [
          "erreur"
        ],
        "title": "ErrorResponse",
        "description": "What every rejection looks like. `erreur` is French: it is displayed to\nthe user verbatim (AUDIT §5.5)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}